more functions more fixes

+++ icadm.py
@@ -4,14 +4,21 @@ from pathlib import Path
 from PIL import Image, ImageChops
 
 from packageland import handler
+
 Image.MAX_IMAGE_PIXELS = None
 
 parser = argparse.ArgumentParser(description="ICADM, Image Channel and Directory Modifier")
-parser.add_argument('--directory', metavar='-D', type=str, help='Initial directory to be processed.', required=True)
-parser.add_argument('--command', metavar='-C', type=str,
-                    help='split, merge, 4split, 4merge, flatten, unflatten, tga_png (tga_png is destructive!)',
+parser.add_argument('--directory', '-d', metavar='-D', type=str, help='Initial directory to be processed.',
                     required=True)
-parser.add_argument('--output', metavar='-O', type=str, help='output image type, png=default or tga.')
+parser.add_argument('--command', '-c', metavar='-C', type=str,
+                    help='split, merge, 4split, 4merge, flatten, unflatten, tga_png (tga_png is destructive!), solid_colors',
+                    required=True)
+parser.add_argument('--output', '-o', metavar='-O', type=str, default='png', help='output image type, png or tga, '
+                                                                                  'default=png')
+parser.add_argument('--parallel', '-p', metavar='-P', action=argparse.BooleanOptionalAction,
+                    help='multicore processing')
+parser.add_argument('--multiplier', '-m', metavar='-M', type=int, default=5,
+                    help='if using multicore processing, job multiplier per core. default = 5')
 args = parser.parse_args()
 
 if args.directory.endswith('/') or args.directory.endswith('\\'):
@@ -20,18 +27,33 @@ else:
     folder = args.directory
 p = Path(folder)
 command = args.command
+parallel = args.parallel
+multiplier = args.multiplier
+output_suffix = args.output.lower()
 
 
-def save(image, path):
+def save(image, path, suffix=None):
     path.parent.mkdir(exist_ok=True, parents=True)
-    if args.output is not None and args.output.lower() == 'tga':
+    if suffix is not None:
+        image.save(path.with_suffix(suffix))
+    elif output_suffix is not None and output_suffix == 'tga':
+        # make this not stupid later
         image.save(path.with_suffix('.tga'))
     else:
         image.save(path.with_suffix('.png'), compress_level=1)
 
 
+def make_path(path, d_suffix, flat=True):
+    out_folder = Path(str(folder) + d_suffix)
+    if flat:
+        out_path = Path.joinpath(out_folder, flatten_path(path))
+    else:
+        out_path = Path.joinpath(out_folder, Path(*path.parts[1:]))
+    return out_path
+
+
 def flatten_path(path):
-    flattened_path = Path(str(path.parent).replace('\\', '') + path.stem + ".png")
+    flattened_path = Path(str(path.parent).replace('\\', '') + path.name)
     return flattened_path
 
 
@@ -40,22 +62,29 @@ def unflatten_path(path):
     return unflattened_path
 
 
+def check_path_exists(path):
+    if not path.exists():
+        if path.with_suffix('.png').exists():
+            path = path.with_suffix('.png')
+        if path.with_suffix('.tga').exists():
+            path = path.with_suffix('.tga')
+    return path
+
+
 def do_thing_flatten(path):
     """
     flatten dir and save to PNG
     """
     with Image.open(path) as image:
-        out_folder = Path(str(folder) + '_RGBA')
-        out_path = Path.joinpath(out_folder, flatten_path(path))
+        out_path = make_path(path, '_RGBA')
         save(image.convert('RGBA'), out_path)
 
 
 def do_thing_unflatten(path):
-    flattened_folder = Path(str(folder) + '_RGBA')
-    flattened_path = Path.joinpath(flattened_folder, flatten_path(path))
+    flattened_path = make_path(path, '_RGBA')
     try:
         with Image.open(flattened_path) as image:
-            merged_path = Path.joinpath(Path('./output'), Path(*path.parts[1:]).with_suffix(".tga"))
+            merged_path = Path.joinpath(Path('./output'), Path(*path.parts[1:]))
             save(image, merged_path)
     except FileNotFoundError:
         pass
@@ -66,12 +95,12 @@ def do_thing_split_RGB_A(path):
     flatten dir, separate RGB and A, and save to PNG
     """
     with Image.open(path) as image:
-        rgb_folder, alpha_folder = [Path(str(folder) + '_RGB'), Path(str(folder) + '_A')]
-        rgb_path = Path.joinpath(rgb_folder, flatten_path(path))
-        alpha_path = Path.joinpath(alpha_folder, flatten_path(path))
+        rgb_path = make_path(path, '_RGB')
+        alpha_path = make_path(path, '_A')
         save(image.convert('RGB'), rgb_path)
         try:
             with image.getchannel('A') as alpha:
+                # check if pure white
                 if ImageChops.invert(alpha).getbbox():
                     save(alpha, alpha_path)
         except ValueError:
@@ -83,13 +112,10 @@ def do_thing_split_R_G_B_A(path):
     flatten dir, separate R, G, B, A, and save to PNG
     """
     with Image.open(path) as image:
-        # todo: first part to function.
-        red_folder, green_folder, blue_folder, alpha_folder = [Path(str(folder) + '_R'), Path(str(folder) + '_G'),
-                                                               Path(str(folder) + '_B'), Path(str(folder) + '_A')]
-        red_path, green_path, blue_path, alpha_path = [Path.joinpath(red_folder, flatten_path(path)),
-                                                       Path.joinpath(green_folder, flatten_path(path)),
-                                                       Path.joinpath(blue_folder, flatten_path(path)),
-                                                       Path.joinpath(alpha_folder, flatten_path(path))]
+        red_path, green_path, blue_path, alpha_path = [make_path(path, '_R'),
+                                                       make_path(path, '_G'),
+                                                       make_path(path, '_B'),
+                                                       make_path(path, '_A')]
         save(image.getchannel('R'), red_path)
         save(image.getchannel('G'), green_path)
         save(image.getchannel('B'), blue_path)
@@ -101,27 +127,44 @@ def do_thing_split_R_G_B_A(path):
             pass
 
 
+def do_thing_get_solid_colors(path):
+    solid_color_path = make_path(path, '_S', False)
+    image = Image.open(path)
+    try:
+        if len(image.getcolors()) == 1:
+            save(image, solid_color_path)
+            image.close()
+            path.unlink()
+        else:
+            image.close()
+    except TypeError:
+        # if color is greater than 256
+        image.close()
+
+
 def do_thing_TGA_PNG(path):
     """
     convert TGA to PNG
     """
     if path.suffix == '.tga':
-        with Image.open(path) as image:
+        image = Image.open(path)
+        if not image.mode == 'RGBA':
             image.save(path.with_suffix('.png'), compress_level=1)
-        path.unlink()
+            image.close()
+            path.unlink()
+        else:
+            pass
 
 
 def do_thing_merge_R_G_B_A(path):
-    red_folder, green_folder, blue_folder, alpha_folder = [Path(str(folder) + '_R'), Path(str(folder) + '_G'),
-                                                           Path(str(folder) + '_B'), Path(str(folder) + '_A')]
-    red_path, green_path, blue_path, alpha_path = [Path.joinpath(red_folder, flatten_path(path)),
-                                                   Path.joinpath(green_folder, flatten_path(path)),
-                                                   Path.joinpath(blue_folder, flatten_path(path)),
-                                                   Path.joinpath(alpha_folder, flatten_path(path))]
+    red_path, green_path, blue_path, alpha_path = [make_path(path, '_R'),
+                                                   make_path(path, '_G'),
+                                                   make_path(path, '_B'),
+                                                   make_path(path, '_A')]
     try:
         with Image.merge('RGB', [Image.open(red_path).convert('L'), Image.open(green_path).convert('L'),
                                  Image.open(blue_path).convert('L')]) as image:
-            merged_path = Path.joinpath(Path('./output'), Path(*path.parts[1:]))
+            merged_path = make_path(path, '_output', False)
             try:
                 with Image.open(alpha_path).convert('L') as alpha:
                     image.putalpha(alpha)
@@ -134,12 +177,13 @@ def do_thing_merge_R_G_B_A(path):
 
 
 def do_thing_merge_RGB_A(path):
-    rgb_folder, alpha_folder = [Path(str(folder) + '_RGB'), Path(str(folder) + '_A')]
-    rgb_path = Path.joinpath(rgb_folder, flatten_path(path))
-    alpha_path = Path.joinpath(alpha_folder, flatten_path(path))
+    rgb_path = make_path(path, '_RGB')
+    rgb_path = check_path_exists(rgb_path)
+    alpha_path = make_path(path, '_A')
+    alpha_path = check_path_exists(alpha_path)
     try:
         with Image.open(rgb_path) as image:
-            merged_path = Path.joinpath(Path('./output'), Path(*path.parts[1:]))
+            merged_path = make_path(path, '_output', False)
             try:
                 with Image.open(alpha_path).convert('L') as alpha:
                     image.putalpha(alpha)
@@ -166,13 +210,14 @@ def read_command(command):
         return do_thing_merge_R_G_B_A
     if "unflatten" == command.lower():
         return do_thing_unflatten
+    if "solid_colors" == command.lower():
+        return do_thing_get_solid_colors
 
 
 if __name__ == '__main__':
     grabber = list(p.glob('**/*.*'))
     function = read_command(command)
-    parallel = True
     if parallel:
-        handler.parallel_process(grabber, function, 5)
+        handler.parallel_process(grabber, function, multiplier)
     else:
         handler.solo_process(grabber, function)
